Move AnimatedButton to core
