--- packages/mobile/src/components/animated-button/index.ts
@@ -1,4 +1,0 @@
-import { AnimatedButtonProps as _AnimatedButtonProps } from './AnimatedButton'
-
-export { AnimatedButton } from './AnimatedButton'
-export type AnimatedButtonProps = _AnimatedButtonProps
