+++ packages/mobile/src/components/bottom-tab-bar/BottomTabBarButton.tsx
@@ -13,7 +13,7 @@ import IconProfileDark from 'app/assets/animations/iconProfileDark.json'
 import IconProfileLight from 'app/assets/animations/iconProfileLight.json'
 import IconTrendingDark from 'app/assets/animations/iconTrendingDark.json'
 import IconTrendingLight from 'app/assets/animations/iconTrendingLight.json'
-import { AnimatedButton } from 'app/components/animated-button'
+import { AnimatedButton } from 'app/components/core'
 import { makeStyles } from 'app/styles'
 import { GestureResponderHandler } from 'app/types/gesture'
 import { useThemeColors } from 'app/utils/theme'
