+++ packages/mobile/src/components/core/index.ts
@@ -1,17 +1,18 @@
+export * from './AnimatedButton'
 export * from './Button'
 export * from './CardList'
 export * from './DynamicImage'
 export * from './EmptyTile'
+export * from './GradientIcon'
+export * from './GradientText'
+export * from './Hyperlink'
 export * from './IconButton'
+export * from './Link'
+export * from './Screen'
+export * from './SearchInput'
+export * from './SegmentedControl'
+export * from './Switch'
+export * from './Text'
 export * from './TextButton'
 export * from './Tile'
 export * from './VirtualizedScrollView'
-export * from './GradientText'
-export * from './GradientIcon'
-export * from './SearchInput'
-export * from './SegmentedControl'
-export * from './Screen'
-export * from './Link'
-export * from './Text'
-export * from './Switch'
-export * from './Hyperlink'
