+++ packages/mobile/src/components/now-playing-drawer/ActionsBar.tsx
@@ -34,8 +34,7 @@ import IconAirplay from 'app/assets/images/iconAirplay.svg'
 import IconChromecast from 'app/assets/images/iconChromecast.svg'
 import IconKebabHorizontal from 'app/assets/images/iconKebabHorizontal.svg'
 import IconShare from 'app/assets/images/iconShare.svg'
-import { AnimatedButton } from 'app/components/animated-button'
-import { IconButton } from 'app/components/core'
+import { AnimatedButton, IconButton } from 'app/components/core'
 import { useDispatchWeb } from 'app/hooks/useDispatchWeb'
 import { useSelectorWeb } from 'app/hooks/useSelectorWeb'
 import { useThemedStyles } from 'app/hooks/useThemedStyles'
