+++ packages/mobile/src/components/now-playing-drawer/AudioControls.tsx
@@ -20,8 +20,7 @@ import IconShuffleOnDark from 'app/assets/animations/iconShuffleOnDark.json'
 import IconShuffleOnLight from 'app/assets/animations/iconShuffleOnLight.json'
 import IconNext from 'app/assets/images/iconNext.svg'
 import IconPrev from 'app/assets/images/iconPrev.svg'
-import { AnimatedButton } from 'app/components/animated-button'
-import { IconButton } from 'app/components/core'
+import { AnimatedButton, IconButton } from 'app/components/core'
 import * as haptics from 'app/haptics'
 import { useDispatchWeb } from 'app/hooks/useDispatchWeb'
 import { useSelectorWeb } from 'app/hooks/useSelectorWeb'
